+++ src/extension.ts
@@ -82,12 +82,20 @@ export function activate(context: vscode.ExtensionContext): void {
     );
   }
 
-  registerEmulatorCommand("emacs-mcx.universalArgumentDigit", (emulator, args) => {
+  registerEmulatorCommand("emacs-mcx.subsequentArgumentDigit", (emulator, args) => {
     const arg = args[0];
     if (typeof arg !== "number") {
       return;
     }
-    emulator.universalArgumentDigit(arg);
+    emulator.subsequentArgumentDigit(arg);
+  });
+
+  registerEmulatorCommand("emacs-mcx.digitArgument", (emulator, args) => {
+    const arg = args[0];
+    if (typeof arg !== "number") {
+      return;
+    }
+    emulator.digitArgument(arg);
   });
 
   registerEmulatorCommand("emacs-mcx.typeChar", (emulator, args) => {
@@ -110,6 +118,22 @@ export function activate(context: vscode.ExtensionContext): void {
 
   registerEmulatorCommand("emacs-mcx.isearchBackward", (emulator) => {
     emulator.runCommand("isearchBackward");
+  });
+
+  registerEmulatorCommand("emacs-mcx.isearchForwardRegexp", (emulator) => {
+    emulator.runCommand("isearchForwardRegexp");
+  });
+
+  registerEmulatorCommand("emacs-mcx.isearchBackwardRegexp", (emulator) => {
+    emulator.runCommand("isearchBackwardRegexp");
+  });
+
+  registerEmulatorCommand("emacs-mcx.queryReplace", (emulator) => {
+    emulator.runCommand("queryReplace");
+  });
+
+  registerEmulatorCommand("emacs-mcx.queryReplaceRegexp", (emulator) => {
+    emulator.runCommand("queryReplaceRegexp");
   });
 
   registerEmulatorCommand("emacs-mcx.isearchAbort", (emulator) => {
@@ -136,6 +160,10 @@ export function activate(context: vscode.ExtensionContext): void {
 
   registerEmulatorCommand("emacs-mcx.universalArgument", (emulator) => {
     emulator.universalArgument();
+  });
+
+  registerEmulatorCommand("emacs-mcx.negativeArgument", (emulator) => {
+    return emulator.negativeArgument();
   });
 
   registerEmulatorCommand("emacs-mcx.killLine", (emulator) => {
@@ -274,13 +302,31 @@ export function activate(context: vscode.ExtensionContext): void {
     emulator.runCommand("paredit.backwardUpSexp");
   });
 
+  registerEmulatorCommand("emacs-mcx.paredit.markSexp", (emulator) => {
+    emulator.runCommand("paredit.markSexp");
+  });
+
   registerEmulatorCommand("emacs-mcx.paredit.killSexp", (emulator) => {
     emulator.runCommand("paredit.killSexp");
+  });
+
+  registerEmulatorCommand("emacs-mcx.paredit.backwardKillSexp", (emulator) => {
+    emulator.runCommand("paredit.backwardKillSexp");
   });
 
   vscode.commands.registerCommand("emacs-mcx.executeCommands", async (...args: any[]) => {
     if (1 <= args.length) {
       executeCommands(args[0]);
+    }
+  });
+
+  registerEmulatorCommand("emacs-mcx.executeCommandWithPrefixArgument", (emulator, args) => {
+    if (typeof args === "object" && args != null) {
+      if ("command" in args) {
+        emulator.executeCommandWithPrefixArgument(args["command"], args["args"], args["prefixArgumentKey"]);
+      }
+    } else if (Array.isArray(args) && args.length >= 1) {
+      emulator.executeCommandWithPrefixArgument(args[0], args[1], args[2]);
     }
   });
 }